UPDATE: editor's pick 댓글 삭제 ajax로 수정

+++ editor_reviews/views.py
@@ -134,10 +134,25 @@ def editor_review_comment(request, pk):
     return JsonResponse(data)
 
 def editor_review_comment_delete(request, reviewpk, commentpk):
-    comment = get_object_or_404(Editor_Review_Comment, pk=commentpk)
-    comment.delete()
+    if request.is_ajax():
+        user = request.user
+        auth_status = {"NL": "notLogined", "NA": "notAnAuthor", "OK": "OK"}
+        comment = get_object_or_404(Editor_Review_Comment, pk=commentpk)
 
-    return redirect('editors_pick:detail', reviewpk)
+        ctx = {
+            "status": auth_status["OK"]
+        }
+
+        if user == AnonymousUser():
+            ctx["status"] = auth_status["NL"]
+        
+        elif comment.author != user:
+            ctx["status"] = auth_status["NA"]
+
+        else:
+            comment.delete()
+
+    return JsonResponse(ctx)
 
 
 @login_required
